+++ Dialog.php
@@ -39,6 +39,13 @@ class Dialog extends Widget
     {
         parent::init();
         echo Html::beginTag('div', $this->options) . "\n";
+        
+        //Fix for closing icon (x) not showing up in dialog
+        $this->getView()->registerJs("
+            var bootstrapButton = $.fn.button.noConflict(); 
+            $.fn.bootstrapBtn = bootstrapButton;", 
+            \yii\web\View::POS_READY
+        );
     }
 
     /**