+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Yii Framework 2 jui extension Change Log
 -----------------------
 
 - Bug #46: `yii\jui\Selectable` add support for `begin()`, `end()` widget methods (fdezmc)
-- Enh: Use `jQuery` instead of `$` in generated code to avoid conflicts (samdark)
+- Enh #56: Use `jQuery` instead of `$` in generated code to avoid conflicts (samdark)
 
 
 2.0.6 July 22, 2016