+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii Framework 2 jui extension Change Log
 
 - Bug #46: `yii\jui\Selectable` add support for `begin()`, `end()` widget methods (fdezmc)
 - Enh #56: Use `jQuery` instead of `$` in generated code to avoid conflicts (samdark)
+- Chg #55: `yii\jui\AutoComplete::run()` now returns output instead of echoing it (unlimix) 
 
 
 2.0.6 July 22, 2016